rename test data files
