+++ projects/_01_ingest/ons_pd/tests/test_ingest_ons_data.py
@@ -3,8 +3,10 @@ from unittest.mock import patch, Mock
 
 import projects._01_ingest.ons_pd.jobs.ingest_ons_data as job
 from utils import utils
-from projects._01_ingest.unittest_data.data import ONSData as Data
-from projects._01_ingest.unittest_data.schemas import ONSData as Schemas
+from projects._01_ingest.unittest_data.ingest_test_file_data import ONSData as Data
+from projects._01_ingest.unittest_data.ingest_test_file_schemas import (
+    ONSData as Schemas,
+)
 
 PATCH_PATH = "projects._01_ingest.ons_pd.jobs.ingest_ons_data"
